requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-socketio==5.1.1
tweepy==4.10.0
instabot==0.117.0
python-dotenv==0.19.2